add silent option
suppress error if silentlycontinue is specified. an example use would be when a ticket # is known, but it is unknown whether it is a service or project ticket; in this case we'd check one area and then the other if the first returned null, but would be uninterested in the "error" status if the first returned null.